separate blake fast deduciton
