+++ stwo_cairo_prover/crates/prover/src/cairo_air/blake/mod.rs
@@ -1,4 +1,3 @@
 pub mod const_columns;
-pub mod deduce_output;
 pub use const_columns::BlakeSigma;
 pub mod air;
