+++ stwo_cairo_prover/crates/prover/src/components/prelude.rs
@@ -21,9 +21,6 @@ pub mod proving {
     pub use stwo_prover::core::poly::circle::{CanonicCoset, CircleEvaluation};
     pub use stwo_prover::core::poly::BitReversedOrder;
 
-    pub use crate::cairo_air::blake::deduce_output::{
-        BlakeG, BlakeRound, BlakeRoundSigma, TripleXor32,
-    };
     pub use crate::cairo_air::poseidon::deduce_output::{
         PackedCube252, PackedPoseidon3PartialRoundsChain, PackedPoseidonFullRoundChain,
         PackedPoseidonRoundKeys,
@@ -31,6 +28,9 @@ pub mod proving {
     pub use crate::cairo_air::preprocessed::Seq;
     pub(crate) use crate::cairo_air::relations;
     pub use crate::components::utils::*;
+    pub use crate::witness::fast_deduction::blake::{
+        BlakeG, BlakeRound, BlakeRoundSigma, TripleXor32,
+    };
 }
 
 pub mod constraint_eval {
