+++ stwo_cairo_prover/crates/prover/src/witness/mod.rs
@@ -1,5 +1,6 @@
 pub mod blake_context;
 pub mod builtins;
 pub mod cairo;
+pub mod fast_deduction;
 pub mod opcodes;
 pub mod range_checks;
